// src/utils/ttrpgstatblocks/bestiary.d.ts
export type Creature5eTools =
	| {
// ...
			type:
				| {
						type: CreatureType;
						/**
						 * Size of the individual creatures that make up the swarm.
						 */
						swarmSize?:
							| 'F'
							| 'D'
							| 'T'
							| 'S'
							| 'M'
							| 'L'
							| 'H'
							| 'G'
							| 'C'
							| 'V';
						tags?: (
							| string
							| {
									tag: string;
									prefix: string;
							  }
						)[];
				  }
				| CreatureType;
			source: string;
			/**
			 * Sub-source text that is shown when hovered.
			 */
			sourceSub?: string;
			otherSources?: OtherSources;
			reprintedAs?: ReprintedAs;
			alignment?: Align[];
			alignmentPrefix?: string;
			ac: AcItem[];
			hp:
				| {
						average: number;
						formula: string;
				  }
				| {
						special?: string;
				  };
			speed: Speed;
			str: number | null;
			dex: number | null;
			con: number | null;
			int: number | null;
			wis: number | null;
			cha: number | null;
			save?: {
				str?: string;
				dex?: string;
				con?: string;
				int?: string;
				wis?: string;
				cha?: string;
			};
			skill?: {
				acrobatics?: string;
				'animal handling'?: string;
				arcana?: string;
				athletics?: string;
				deception?: string;
				history?: string;
				insight?: string;
				intimidation?: string;
				investigation?: string;
				medicine?: string;
				nature?: string;
				perception?: string;
				performance?: string;
				persuasion?: string;
				religion?: string;
				'sleight of hand'?: string;
				stealth?: string;
				survival?: string;
				other?: {
					oneOf?: {
						acrobatics?: string;
						'animal handling'?: string;
						arcana?: string;
						athletics?: string;
						deception?: string;
						history?: string;
						insight?: string;
						intimidation?: string;
						investigation?: string;
						medicine?: string;
						nature?: string;
						perception?: string;
						performance?: string;
						persuasion?: string;
						religion?: string;
						'sleight of hand'?: string;
						stealth?: string;
						survival?: string;
						[k: string]: unknown;
					};
					[k: string]: unknown;
				}[];
			};
			senses?: string[] | null;
			passive: number | string | null;
			languages?: string[] | null;
			pbNote?: string;
			cr?:
				| string
				| {
						cr: string;
						lair?: string;
						coven?: string;
						xp?: number;
				  };
			vulnerable?: DamageVulnerabilityArray;
			resist?: DamageResistArray;
			immune?: DamageImmunityArray;
			conditionImmune?: ConditionImmunityArray;
// ...
	  };
/**
 * F: Fine
 *  D: Diminutive
 *  T: Tiny
 *  S: Small
 *  M: Medium
 *  L: Large
 *  H: Huge
 *  G: Gargantuan
 *  C: Colossal
 *  V: Varies.
 */
export type Size = 'F' | 'D' | 'T' | 'S' | 'M' | 'L' | 'H' | 'G' | 'C' | 'V';
export type CreatureType =
	| 'aberration'
	| 'beast'
	| 'celestial'
	| 'construct'
	| 'dragon'
	| 'elemental'
	| 'fey'
	| 'fiend'
	| 'giant'
	| 'humanoid'
	| 'monstrosity'
	| 'ooze'
	| 'plant'
	| 'undead';
export type Page = number;
export type OtherSources = {
	source: string;
	page?: Page;
}[];
export type ReprintedAs = (
	| string
	| {
			/**
			 * A UID, e.g. "longsword|phb"
			 */
			uid: string;
			/**
			 * A tag used in the renderer, e.g. "creature"
			 */
			tag: string;
	  }
)[];
export type Align =
	| Alignment
	| {
			alignment: Alignment[];
			chance?: number;
			note?: string;
	  }
	| {
			special: string;
	  };
/**
 * L: Lawful
 *  N: Neutral
 *  NX: Neutral (law/chaos axis)
 *  NY: Neutral (good/evil axis)
 *  C: Chaotic
 *  G: Good
 *  E: Evil
 *  U: Unaligned
 *  A: Any
 */
export type Alignment = 'L' | 'N' | 'NX' | 'NY' | 'C' | 'G' | 'E' | 'U' | 'A';
export type AcItem =
	| {
			ac: number;
			from?: string[];
			condition?: string;
			braces?: boolean;
	  }
	| {
			special: string;
	  }
	| number;
// ...
export type _SpeedVal =
	| {
			number: number;
			condition: string;
	  }
	| number;
export type DamageVulnerabilityArray =
	| (
			| DataDamageType
			| {
					special: string;
			  }
			| {
					preNote?: string;
					vulnerable: DamageVulnerabilityArray;
					note?: string;
					cond?: boolean;
			  }
	  )[]
	| null;
export type DataDamageType =
	| 'acid'
	| 'bludgeoning'
	| 'cold'
	| 'fire'
	| 'force'
	| 'lightning'
	| 'necrotic'
	| 'piercing'
	| 'poison'
	| 'psychic'
	| 'radiant'
	| 'slashing'
	| 'thunder';
export type DamageResistArray =
	| (
			| DataDamageType
			| {
					special: string;
			  }
			| {
					preNote?: string;
					resist: DamageResistArray;
					note?: string;
					cond?: boolean;
			  }
	  )[]
	| null;
export type DamageImmunityArray =
	| (
			| DataDamageType
			| {
					special: string;
			  }
			| {
					preNote?: string;
					immune: DamageImmunityArray;
					note?: string;
					cond?: boolean;
			  }
	  )[]
	| null;
export type ConditionImmunityArray =
	| (
			| DataCondition
			| {
					special: string;
			  }
			| {
					preNote?: string;
					conditionImmune: ConditionImmunityArray;
					note?: string;
			  }
	  )[]
	| null;
export type DataCondition =
	| 'blinded'
	| 'charmed'
	| 'deafened'
	| 'exhaustion'
	| 'frightened'
	| 'grappled'
	| 'incapacitated'
	| 'invisible'
	| 'paralyzed'
	| 'petrified'
	| 'poisoned'
	| 'prone'
	| 'restrained'
	| 'stunned'
	| 'unconscious'
	| 'disease';
export type Srd = boolean | string;
export type BasicRules = boolean;
export type _ArrayOfSpell = (
	| string
	| {
			entry: string;
			hidden: boolean;
			[k: string]: unknown;
	  }
)[];
// ...
export interface BestiarySchema {
	monster?: Creature[];
	_meta?: MetaBlock;
}
export interface EntrySpellcasting {
	name: string;
	type: string;
	source?: string;
	page?: Page;
	srd?: Srd;
	basicRules?: BasicRules;
	/**
	 * A generic object for storing special data for external use-cases. Keys prefixed with "rd-" should be added as "data-" HTML attributes when rendering to HTML.
	 */
	data?: {
		[k: string]: unknown;
	};
	id?: string;
	headerEntries?: BestiarySchema[];
	constant?: _ArrayOfSpell;
	will?: _ArrayOfSpell;
	ritual?: _ArrayOfSpell;
	rest?: EntrySpellcastingFrequency;
	daily?: EntrySpellcastingFrequency;
	weekly?: EntrySpellcastingFrequency;
	yearly?: EntrySpellcastingFrequency;
	charges?: EntrySpellcastingFrequency;
	spells?: {
		'0'?: {
			spells: string[];
		};
		'1'?: EntrySpellcastingLevel1To9;
		'2'?: EntrySpellcastingLevel1To9;
		'3'?: EntrySpellcastingLevel1To9;
		'4'?: EntrySpellcastingLevel1To9;
		'5'?: EntrySpellcastingLevel1To9;
		'6'?: EntrySpellcastingLevel1To9;
		'7'?: EntrySpellcastingLevel1To9;
		'8'?: EntrySpellcastingLevel1To9;
		'9'?: EntrySpellcastingLevel1To9;
	};
	/**
	 * Allows the above properties to be specified, but not rendered. Useful if e.g. a creature can only cast one spell, and this is rendered in the header line.
	 */
	hidden?: (
		| 'constant'
		| 'will'
		| 'rest'
		| 'daily'
		| 'weekly'
		| 'yearly'
		| 'ritual'
		| 'spells'
	)[];
	footerEntries?: string[];
	ability?: 'str' | 'dex' | 'con' | 'int' | 'wis' | 'cha';
	/**
	 * Implicitly "trait"
	 */
	displayAs?: 'trait' | 'action';
	/**
	 * The UID of an item which provides the charges required to cast any "charges" spells.
	 */
	chargesItem?: string;
}
export interface EntrySpellcastingFrequency {
	'1'?: _ArrayOfSpell;
	'2'?: _ArrayOfSpell;
	'3'?: _ArrayOfSpell;
	'4'?: _ArrayOfSpell;
	'5'?: _ArrayOfSpell;
	'6'?: _ArrayOfSpell;
	'7'?: _ArrayOfSpell;
	'8'?: _ArrayOfSpell;
	'9'?: _ArrayOfSpell;
	'9e'?: _ArrayOfSpell;
	'8e'?: _ArrayOfSpell;
	'7e'?: _ArrayOfSpell;
	'6e'?: _ArrayOfSpell;
	'5e'?: _ArrayOfSpell;
	'4e'?: _ArrayOfSpell;
	'3e'?: _ArrayOfSpell;
	'2e'?: _ArrayOfSpell;
	'1e'?: _ArrayOfSpell;
}
export interface EntrySpellcastingLevel1To9 {
	lower?: number;
	slots?: number;
	spells: string[];
}
// ...
export interface MetaBlock {
	dependencies?: {
		monster?: string[];
		[k: string]: unknown;
	};
	otherSources?: {
		/**
		 * Keys are other sources to be loaded; values are `otherSources` sources from that source to search for.
		 */
		monster?: {
			[k: string]: unknown;
		};
		[k: string]: unknown;
	};
	[k: string]: unknown;
}